requirements.txt
apache-beam==2.69.0
attrs==25.4.0
beartype==0.21.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crcmod==1.7
cryptography==46.0.3
dnspython==2.8.0
docopt==0.6.2
fastavro==1.12.1
fasteners==0.20
grpcio==1.76.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.11
jsonpickle==3.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python==2.2.15
numpy==2.2.6
objsize==0.7.1
orjson==3.11.4
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==18.1.0
pyarrow-hotfix==0.7
pycparser==2.23
pydot==1.4.2
PyJWT==2.10.1
pymongo==4.15.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.15.0
urllib3==2.5.0
zstandard==0.25.0
